# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DIDEE Digital Flagship API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: collectionSlug
          in: query
          schema: { type: string }
        - name: categorySlug
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: isNew
          in: query
          schema: { type: boolean }
        - name: isBestSeller
          in: query
          schema: { type: boolean }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: sortBy
          in: query
          schema: { type: string, enum: [price_asc, price_desc, newest, name] }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                required: [products, total]
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
                  total: { type: integer }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/new-arrivals:
    get:
      operationId: getNewArrivals
      tags: [products]
      summary: Get newest products
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 8 }
      responses:
        "200":
          description: New arrivals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
  /products/best-sellers:
    get:
      operationId: getBestSellers
      tags: [products]
      summary: Get best-selling products
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 8 }
      responses:
        "200":
          description: Best sellers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /products/{slug}/variants:
    get:
      operationId: listProductVariants
      tags: [products]
      summary: List variants for a product
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Variants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProductVariant" }
    post:
      operationId: createProductVariant
      tags: [products]
      summary: Add a variant to a product
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductVariantInput"
      responses:
        "201":
          description: Created variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
  /products/{slug}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }

  # ─── Collections ─────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: productSlug
          in: query
          schema: { type: string }
        - name: approved
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a product review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Approve or update a review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── Journal ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, images, status]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        shortDescription: { type: ["string", "null"] }
        price: { type: number }
        comparePrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        collectionSlug: { type: ["string", "null"] }
        categorySlug: { type: ["string", "null"] }
        status: { type: string, enum: [active, draft, archived] }
        featured: { type: boolean }
        isNew: { type: boolean }
        isBestSeller: { type: boolean }
        totalStock: { type: integer }
        avgRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        createdAt: { type: string }
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            variants:
              type: array
              items: { $ref: "#/components/schemas/ProductVariant" }
            reviews:
              type: array
              items: { $ref: "#/components/schemas/Review" }
            collection: { $ref: "#/components/schemas/Collection" }
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name: { type: string }
        description: { type: string }
        shortDescription: { type: string }
        price: { type: number }
        comparePrice: { type: number }
        images: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        collectionSlug: { type: string }
        categorySlug: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        featured: { type: boolean }
        isNew: { type: boolean }
        isBestSeller: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        shortDescription: { type: string }
        price: { type: number }
        comparePrice: { type: number }
        images: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        collectionSlug: { type: string }
        categorySlug: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        featured: { type: boolean }
        isNew: { type: boolean }
        isBestSeller: { type: boolean }
    ProductVariant:
      type: object
      required: [id, productId, sku, stock, price]
      properties:
        id: { type: integer }
        productId: { type: integer }
        size: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        sku: { type: string }
        stock: { type: integer }
        price: { type: number }
    ProductVariantInput:
      type: object
      required: [sku, stock, price]
      properties:
        size: { type: string }
        color: { type: string }
        sku: { type: string }
        stock: { type: integer }
        price: { type: number }

    # ─── Collection ─────────────────────────────────────────────────────────────
    Collection:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        image: { type: ["string", "null"] }
        bannerImage: { type: ["string", "null"] }
        featured: { type: boolean }
        season: { type: ["string", "null"] }
        productCount: { type: integer }
        sortOrder: { type: integer }
...
    Review:
      type: object
      required: [id, productId, rating, approved]
      properties:
        id: { type: integer }
        productId: { type: integer }
        customerName: { type: string }
        rating: { type: integer }
        title: { type: ["string", "null"] }
        body: { type: ["string", "null"] }
        approved: { type: boolean }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [productId, customerName, rating]
      properties:
        productId: { type: integer }
        customerName: { type: string }
        rating: { type: integer }
        title: { type: string }
        body: { type: string }
    ReviewUpdate:
      type: object
      properties:
        approved: { type: boolean }
        title: { type: string }
        body: { type: string }

    # ─── Journal ────────────────────────────────────────────────────────────────
...